[world] Serve static files
Completed Module 2, Part 3 "Serving Files in ASP.NET 5".

The 'Microsoft.AspNet.StaticFiles' library (?) was added to help
serve the static file index.html. Supposedly this is necessary when
hosting the application outside of IIS.

The Startup#Configure method was modified to setup serving of
static content. The order of the Use methods matters, as that sets
up the chain of how content is served.