[contact] Add a 'mail' service
The submitted contact form will be 'mailed'. A mail service interface was
added with just one implementation - logging some info to the Debug
console. The one implemenation is injected into the AppController through
the constructor. This is done with the IServiceCollection in Startup.cs -
a lot like the Java config of Spring