[configs] Use IOptions for configs
IOptions<T> looks to be the preferred way of passing configurations in
ASP.NET 5. Following the instructions from
http://docs.asp.net/en/latest/fundamentals/configuration.html, the
AppSettings class was added to hold the read in configs.

The config.json was reorganized to get the value to the right spot.
Otherwise, the type T would have required an AppSettings field. This is
great for loading configs at app startup, but there are some use cases for
reloading configs on demand - that'll be explored later on.

IOptions is another opt-in bit. The OptionsModel module was added to the
dependency.

It also turns out that the JSON parser in .NET accepts comments!