[db] Complete 'Using the DbContext'
The DbContext is now injected into AppController for querying Trips.
Luckily there's a test DB included with Visual Studio to bang against.

The connection for that is encoded into config.json. The 'extra
initiative' of using IOptions<T> meant that a DBSettings class was
necessary to get things linked up.

The application now has a database to connect to - but there's no actual
tables, data, etc. to read. Seeding the DB is the next lesson!