[db] Hide EF behind the Repo pattern
Doing raw EF calls in the controllers is bad practice. All of the data
access should be hidden in yet another layer, the IWorldRepository
interface. Production will use WorldRepository, tests could use a dummy
implementation.

The Trips is also displayed to the Index view. Some backtracking is
necessary to replicate the skipped Bootstrap sections to make it purty.