[app] Add logging
The logging module was added and registered. So far it just logs to the
debug console, but logging either to a file, Windows logger, or whatever
the standard 'Windows' way would be nice.

One annoying thing is the logger is injected via the constructor. What
would be nice is to find a log4j equivalent - where the logger is just a
private static field that's assigned in a static block
  (e.g. Java: private final static Logger = Logger.getLogger(Foo.class) )