[api] Initial Trip REST-ish controller
Again, a lot like Spring MVC so far - especially with TripController and
AppController having the same structure. The only difference between
getting HTML back vs. raw JSON is the return type (IActionResult vs.
JsonResult).

The funky CamelCase resolver thingy in Startup.cs is so that the
serialized JSON has the typical camelCase structure. Otherwise the JSON
retains .NET's Casing.

The application is also no longer launching the browser on startup. This
is to make it easier to test with other tools, such as Postman