[api] Auto convert Trip <-> TripViewModel
Trip is the class for going between our code and the DB. TripViewModel is
the class for going between our code and the outside world. The AutoMapper
library handles the conversion between the two automatically - with some
prodding in Startup.cs. What's even nicer is that support for common
collection types (e.g. IEnumerable) is included.

One flaw is that some fields of Trip (i.e. Field) are not automatically
mapped. That'll come in the next lesson or two.