[api] Stop: Added Stop controller
A new controller is added for querying just stops. Because a Stop only
makes sense in context of a trip, the URI is rooted in a trip.

The trip is specified by {tripName}. The trip name is automatically
injected into the controller method via a parameter of the same name. With
Spring this would have been done with @PathVariable("tripName").

The WorldRepo is extended with a GetTripByName method.

The introduction of the StopViewModel allows for easy translation for the
API results, and also allows for Stops to be included with serialized
TripViewModels.